requirements.txt
boto3==1.9.239
botocore==1.12.239
certifi==2019.9.11
chardet==3.0.4
Click==7.0
docutils==0.15.2
Flask==1.1.1
Flask-Session==0.3.1
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
Pillow==6.1.0
psycopg2==2.8.3
python-dateutil==2.8.0
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
SQLAlchemy==1.3.10
urllib3==1.25.6
Werkzeug==0.16.0
WTForms==2.2.1
gunicorn